86 auth firebase (#87)
* refactor to use firebase auth

* refactor to use firebase auth

* add firebase sdk web

* add firebase config

* stub out signup form

* refactor to use firebase auth

* refactor to use firebase auth

* refactor to use firebase auth

* add route for signup view

* refactor for firebase auth

* cleanup

* add user guid setter, fix some state change pointers

* cleanup

* hide login button until refactor behavior

* refactor ProtectedRoute to Route

* cleanup

* cleanup

* cleanup

* hide auth button until refactor for login and signup behavior

* cleanup

* stub out the form and handlers

* cleanup

* add default route to jobs comp

* add test link to login

* add route to login

* rename getUserGuideReq

* update logout for firebase

* update logout for firebase

* add field for fb admin key

* add field for fb admin key

* add firebase admin lib

* cleanup comment auth0

* start to refactor to use firebase auth for api

* add fbIdToken param to api calls

* fix user.getIdToken() call to use promise

* fix user.getIdToken()

* update BE api call for fb auth

* update BE api call for fb auth

* update BE api call for fb auth for company route

* update BE api call for fb auth for company route

* refactor jobs api to use fb admin auth

* refactor jobs api call to use fb admin auth

* refactor jobs api call to use fb admin auth

* refactor jobs api call to use fb admin auth

* refactor jobs api call to use fb admin auth

* refactor route call to use fb auth

* fix BE post calls with content-type header

* cleanup

* cleanup

* cleanup

* cleanup

* refactor get to use post for passing fbIdToken

* refactor get to use post for passing fbIdToken

* refactor get to use post for passing fbIdToken

* add logic to avoid initializing fb admin multiple times

* cleanup ref to auth0

* remove ref to auth0

* update config ref for auth0 removal

* update sample config files for removal of auth0

* remove auth0 lib

* cleanup

* remove jwt related packages that were used with auth0

* cleanup

* cleanup

* refactor apis to use firebase auth

* refactor apis to use firebase auth

* refactor apis to use firebase auth

+++ client/src/components/Events/Events.js
@@ -2,11 +2,12 @@ import React from "react";
 import * as ReactDOM from "react-dom";
 import {useParams, Link} from 'react-router-dom';
 import {useState, useEffect} from 'react';
-import { useAuth0 } from "@auth0/auth0-react";
 import { Calendar, momentLocalizer } from 'react-big-calendar';
 import moment from 'moment';
+import { getAuth } from "firebase/auth";
+
 import "react-big-calendar/lib/css/react-big-calendar.css";
-import {useAppAuth} from '../../context/auth-context';
+import {useAuth} from '../../context/auth-context';
 import {useJobs} from '../../context/jobs-context';
 import {useCompany} from '../../context/company-context';
 import { useAlert, ADD } from '../../context/alert-context';
@@ -37,33 +38,48 @@ function Events(props) {
   const [filterEventTakeHomeTestScheduled, setFilterEventTakeHomeTestScheduled] = useState(true);
   const [filterEventTakeHomeTestUnscheduled, setFilterEventTakeHomeTestUnscheduled] = useState(true);
 
-  const { isAuthenticated, user, getAccessTokenSilently } = useAuth0();
-  const { name, picture, email } = user;
-  const {login, getUserGuid, userGuid, userEmail, sessionToken} = useAppAuth();
+  const {login, getUserGuidReq} = useAuth();
   const {jobsDict, getJobsForUser} = useJobs();
   const {companyDict, getCompanies} = useCompany();
   const { alertDispatch } = useAlert();
 
+  const auth = getAuth();
+  const user = auth.currentUser;
+  let userEmail;
+  if (user) {
+    userEmail = user.email;
+  }
+
   const buttonOnClickHandler = (ev) => {
     ev.preventDefault();
-    let {id, name} = ev.target;
-    let eventGuid = name;
-    fetch(`http://localhost:3000/api/events/${eventGuid}`, {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json',
-        'Authorization': `Bearer ${sessionToken}`
-      }
-    })
-    .then(resp => resp.json())
-    .then(json => {
-      if (json.status === 'success') {
-        let copyEventDict = {...eventsDict};
-        delete copyEventDict[eventGuid];
-        setEventsDict(copyEventDict);
-      }
-    })
-    .catch(err => console.error('err', err))
+
+    if (user) {
+      user.getIdToken()
+        .then(fbIdToken => {
+          let {id, name} = ev.target;
+          let eventGuid = name;
+          let body = {fbIdToken}
+          fetch(`http://localhost:3000/api/events/${eventGuid}`, {
+            method: 'DELETE',
+            headers: {
+              'Content-Type': 'application/json'
+            },
+            body: JSON.stringify(body)
+          })
+            .then(resp => resp.json())
+            .then(json => {
+              if (json.status === 'success') {
+                let copyEventDict = {...eventsDict};
+                delete copyEventDict[eventGuid];
+                setEventsDict(copyEventDict);
+              }
+            })
+            .catch(err => console.error('err', err))
+
+        })
+        .catch(err => console.error('err', err))
+
+    }
   }
 
   const inputOnChangeHandler = (ev) => {
@@ -87,45 +103,59 @@ function Events(props) {
     }
   }
 
-  const callSecureApi = async (uGuid) => {
+  const callSecureApi = (uGuid, fbIdToken) => {
     try {
-      const token = await getAccessTokenSilently();
-      login({userEmail: email, sessionToken: token, userGuid: uGuid})
-
-      let eventsUrl;
-      if (!jobId && uGuid) {
-        eventsUrl = `http://localhost:3000/api/events/user/${uGuid}`;
-      } else if (jobId) {
-        eventsUrl = `http://localhost:3000/api/events/job/${jobId}`
-      }
-      if (eventsUrl) {
-        fetch(eventsUrl, {
-          headers: {Authorization: `Bearer ${token}`}
-        })
-          .then(resp => {
-            if (!resp.ok) {
-              alertDispatch({ type: ADD, payload: {type: 'error', message: `HTTP Status Code: ${resp.status}`} })
-            } 
-            return resp.json();
+      if (fbIdToken) {
+        
+        if (userEmail) {
+          login({userEmail, fbIdToken})
+        }
+  
+        let eventsUrl;
+        let body = {fbIdToken}
+        if (!jobId && uGuid) {
+          eventsUrl = `http://localhost:3000/api/events/user/${uGuid}`;
+        } else if (jobId) {
+          eventsUrl = `http://localhost:3000/api/events/job/${jobId}`
+        }
+        if (eventsUrl) {
+          fetch(eventsUrl, {
+            method: 'POST',
+            headers: {
+              'Content-Type': 'application/json'
+            },          
+            body: JSON.stringify(body)
           })
-          .then(json => {
-            if (json.events.length) {
-              let evDict = json.events ? getDictFromAr(json.events): {};
-              setEventsDict(evDict);
-            } else if (json.type === 'error') {
-              alertDispatch({ type: ADD, payload: {type: 'error', message: json.message} })
-            }
-          })
-          .catch(err => console.error('error', err))
-      }
+            .then(resp => {
+              if (!resp.ok) {
+                alertDispatch({ type: ADD, payload: {type: 'error', message: `HTTP Status Code: ${resp.status}`} })
+              } 
+              return resp.json();
+            })
+            .then(json => {
+              if (json.events.length) {
+                let evDict = json.events ? getDictFromAr(json.events): {};
+                setEventsDict(evDict);
+              } else if (json.type === 'error') {
+                alertDispatch({ type: ADD, payload: {type: 'error', message: json.message} })
+              }
+            })
+            .catch(err => console.error('error', err))
+        }
+  
+        if (uGuid) {
+          let companiesUrl = `http://localhost:3000/api/company/all/${uGuid}`;
+          getCompanies({url: companiesUrl, fbIdToken});
+            
+          let jobsUrl = `http://localhost:3000/api/jobs/all/${uGuid}`;
+          getJobsForUser({url: jobsUrl, fbIdToken});
 
-      if (uGuid) {
-        let jobsUrl = `http://localhost:3000/api/jobs/all/${uGuid}`;
-        getJobsForUser({url: jobsUrl, token});
+        }            
 
-        let companiesUrl = `http://localhost:3000/api/company/all/${uGuid}`;
-        getCompanies({url: companiesUrl, token});
+
       }
+      //  })
+      //  .catch(error => console.error('err', error))
 
     } catch (error) {
       console.error('error', error)
@@ -156,10 +186,20 @@ function Events(props) {
   }
 
   useEffect(() => {
-    getUserGuid({userEmail: email})
-      .then(uGuid => {
-        callSecureApi(uGuid);
-      })
+    if (userEmail) {
+      user.getIdToken()
+        .then(fbIdToken => {
+          if (fbIdToken) {
+            getUserGuidReq({userEmail, fbIdToken})
+              .then(uGuid => {
+                
+                callSecureApi(uGuid, fbIdToken);
+              })
+              .catch(error => console.error('err', error))
+          }
+        })
+        .catch(error => console.error('err', error))
+    }
   }, [])
 
   let eventsAr = Object.keys(eventsDict).length ? getArFromDict(eventsDict) : [];
